feat(render): add ext option for custom file extenstion, close #340
